+++ src/main/java/com/kshrd/jpa_hibernate_homework/service/impl/ProductServiceImpl.java
@@ -33,7 +33,7 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public List<Product> getProductByQuantity(Long quantity) {
+    public List<Product> getProductByQuantity(Integer quantity) {
         return null;
     }
